Change sls_price type from DECIMAL to INT

+++ scripts/bronze/ddl_bronze.sql
@@ -52,7 +52,7 @@ CREATE TABLE bronze.crm_sales_details (
 	sls_due_dt INT,
 	sls_sales INT,
 	sls_quantity INT,
-	sls_price DECIMAL(18, 2)
+	sls_price INT
 );
 GO
